+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 ## Added
 
+- Action "repeatLastAction" to repeat the last called action or mapstring
+- Mapping to execute the last executed command again (requires commandhist to be on "useronly" or "all")
 - Cheatsheet for default mappings and general usage of Vieb to the help page, website and readme
 - Example config for Surfingkeys, Pentadactyl and Vim Vixen
 - Support for numpad/keypad keys in mappings, such as k3, kEnd or kPlus